Respawn and Crossfade Update
Press R when in-game to respawn. It isn't a permanent input, it's just to show off the super duper epic amazing crossfade I made. I programmed it so that you can reference a scene to load from other scripts as well. I can explain this better when we talk next time! :)